+++ ui/app/src/components/ContentTypeManagement/components/TypeDetailsViewHeader.tsx
@@ -25,6 +25,7 @@ import Typography from '@mui/material/Typography';
 import { ItemTypeIcon } from '../../ItemTypeIcon';
 import { FormattedMessage } from 'react-intl';
 import React from 'react';
+import { createComponentId } from '../../../utils/system';
 
 export type TypeDetailsHeaderActionTarget = 'properties' | 'template' | 'jsController' | 'groovyController' | 'deleted';
 
@@ -38,7 +39,7 @@ export function TypeDetailsViewHeader({ type, onActionClick }: TypeDetailsViewHe
 	const handleDeleteType: ButtonProps['onClick'] = (e) => {
 		dispatch(
 			pushDialog({
-				component: 'craftercms.components.DeleteContentTypeDialog',
+				component: createComponentId('DeleteContentTypeDialog'),
 				props: {
 					contentType: type,
 					onComplete() {